Add parser class
